requirements.txt
fastapi==0.110.0
numpy==1.22.3
Pillow==10.2.0
redis==3.5.3
tensorflow==2.12.0
uvicorn==0.20.0
python-multipart==0.0.9
python-dotenv==0.21.0
gunicorn
